move view handlers to view class
